Add files via upload
An application where information of a student can be added, such as name, department, CGPA, sex, marital status, country. These information can be sorted according to different fields. Information can be added, updated or removed. There is also a search option.
There is a problem with the GUI that the alignment of the boxes in which users input their information is not perfect.